exp_files/requirements.txt
# Wake Word Detection
pvporcupine
pyaudio

# Speech-to-Text (Vosk)
vosk
sounddevice

# Text-to-Speech (Piper)
piper-tts
